+++ packages/desktop_autopaste/lib/src/ffi/desktop_autopaste_client.dart
@@ -2,7 +2,10 @@ import '../focused_text_edit_operation.dart';
 import '../focused_text_field_context.dart';
 
 abstract interface class DesktopAutopasteClient {
-  Future<bool> pasteIntoCursorViaClipboard(String text);
+  Future<bool> pasteIntoCursorViaClipboard(
+    String text, {
+    required Duration prePasteDelay,
+  });
 
   Future<FocusedTextFieldContext> getFocusedTextFieldContext({
     int? maxCharsBefore,